requirements.txt
python-telegram-bot[asyncio]==21.5
pyyaml==6.0.2
